Move factorial2 tests to test_overlap.py
